+++ lib/src/country_parser.dart
@@ -15,6 +15,7 @@ import 'res/strings/ku.dart';
 import 'res/strings/lt.dart';
 import 'res/strings/lv.dart';
 import 'res/strings/nb.dart';
+import 'res/strings/nl.dart';
 import 'res/strings/nn.dart';
 import 'res/strings/np.dart';
 import 'res/strings/pl.dart';
@@ -48,12 +49,30 @@ class CountryParser {
     return _getFromCode(countryCode.toUpperCase());
   }
 
+  /// Returns a single country if it matches the given [phoneCode] (e164_cc).
+  ///
+  /// Throws a [StateError] if no matching element is found.
+  static Country parsePhoneCode(String phoneCode) {
+    return _getFromPhoneCode(phoneCode);
+  }
+
   /// Returns a single country that matches the given [countryCode] (iso2_cc).
   ///
   /// Returns null if no matching element is found.
   static Country? tryParseCountryCode(String countryCode) {
     try {
       return parseCountryCode(countryCode);
+    } catch (_) {
+      return null;
+    }
+  }
+
+  /// Returns a single country that matches the given [phoneCode] (e164_cc).
+  ///
+  /// Returns null if no matching element is found.
+  static Country? tryParsePhoneCode(String phoneCode) {
+    try {
+      return parsePhoneCode(phoneCode);
     } catch (_) {
       return null;
     }
@@ -101,6 +120,15 @@ class CountryParser {
     } catch (_) {
       return null;
     }
+  }
+
+  /// Returns a country that matches the [countryCode] (e164_cc).
+  static Country _getFromPhoneCode(String phoneCode) {
+    return Country.from(
+      json: countryCodes.singleWhere(
+        (Map<String, dynamic> c) => c['e164_cc'] == phoneCode,
+      ),
+    );
   }
 
   /// Returns a country that matches the [countryCode] (iso2_cc).
@@ -223,6 +251,8 @@ class CountryParser {
         return lv;
       case 'lt':
         return lt;
+      case 'nl':
+        return nl;
       case 'en':
       default:
         return en;
@@ -256,6 +286,7 @@ class CountryParser {
       const Locale('de'),
       const Locale('lv'),
       const Locale('lv'),
+      const Locale('nl'),
       const Locale.fromSubtags(languageCode: 'zh', scriptCode: 'Hans'),
       const Locale.fromSubtags(languageCode: 'zh', scriptCode: 'Hant'),
     ]..removeWhere((Locale l) => exclude.contains(l));