+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.0.19
+  * Add support for Dutch localization
+  * Add `parsePhoneCode` and `tryParsePhoneCode`
+
+## 2.0.18
+  * Fix Hindi Translations
+  * Removed old country codes for Kosovo (+381 & +386)
+
 ## 2.0.17
   * Add option to hide search bar
     ```Dart